enhance: Remove not in used BinlogIterator
